Fix signalr alert cycle, alert state machine, email notifications

+++ energy-backend.Application/Models/DeviceRequestDto.cs
@@ -5,5 +5,5 @@ public class DeviceRequestDto
     public Guid OrganisationId { get; set; }
     public string Name { get; set; } = string.Empty;
     public string Type { get; set; } = string.Empty;
-    public float RatedPowerWatts { get; set; }
+    public double RatedPowerWatts { get; set; }
 }
