+++ energy-backend.Core/Interfaces/IAlertEvaluationRepository.cs
@@ -2,7 +2,7 @@ namespace energy_backend.Core.Interfaces;
 
 public interface IAlertEvaluationRepository
 {
-    Task<(DateTime? LatestTimestamp, float TotalWatts)> GetLatestMinutePowerSumAsync(
+    Task<(DateTime? LatestTimestamp, double TotalWatts)> GetLatestMinutePowerSumAsync(
         Guid orgId,
         CancellationToken ct = default);
 }
